src/requirements.txt
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.2
opencv-python==4.8.0.76
easyocr==1.7.0
pymupdf==1.23.7
Pillow==10.0.1

